+++ website/app/docs/guides/_meta.js
@@ -1,3 +1,0 @@
-export default {
-  "static-generation": "Static Generation",
-};